Refactor separating api endpoints from the ones for static files 🌉
