--- src/app/controllers/api.controller.ts
@@ -1,10 +1,0 @@
-import { Context, Get, HttpResponseOK } from '@foal/core';
-
-export class ApiController {
-
-  @Get('/')
-  index(ctx: Context) {
-    return new HttpResponseOK('Hello world!');
-  }
-
-}
