+++ src/app/controllers/api/api.controller.ts
@@ -1,0 +1,10 @@
+import { controller } from "@foal/core";
+import { RecipeApiController } from "./recipe.controller";
+import { SearchApiController } from "./search.controller";
+
+export class ApiController {
+  subControllers = [
+    controller("/recipe", RecipeApiController),
+    controller("/search", SearchApiController),
+  ];
+}
