--- src/app/controllers/search.controller.spec.ts
@@ -1,30 +1,0 @@
-// std
-import { ok, strictEqual } from 'assert';
-
-// 3p
-import { Context, createController, getHttpMethod, getPath, isHttpResponseOK } from '@foal/core';
-
-// App
-import { SearchController } from './search.controller';
-
-describe('SearchController', () => {
-
-  let controller: SearchController;
-
-  beforeEach(() => controller = createController(SearchController));
-
-  describe('has a "foo" method that', () => {
-
-    it('should handle requests at GET /.', () => {
-      strictEqual(getHttpMethod(SearchController, 'foo'), 'GET');
-      strictEqual(getPath(SearchController, 'foo'), '/');
-    });
-
-    it('should return an HttpResponseOK.', () => {
-      const ctx = new Context({});
-      ok(isHttpResponseOK(controller.foo(ctx)));
-    });
-
-  });
-
-});
