+++ src/app/controllers/static/pages.controller.ts
@@ -1,0 +1,19 @@
+import { Context, Get, HttpResponseNotFound, render } from "@foal/core";
+
+export class PagesController {
+  @Get("recipe/:recipeId")
+  async getRecipePage(ctx: Context) {
+    if (!ctx.request.accepts("html")) {
+      return new HttpResponseNotFound();
+    }
+    return render("./public/pages/recipe.html");
+  }
+
+  @Get("search")
+  async serachPage(ctx: Context) {
+    if (!ctx.request.accepts("html")) {
+      return new HttpResponseNotFound();
+    }
+    return render("./public/pages/search.html");
+  }
+}
